refactor: 모델 내에 선언된 열거형 분리 및 개행 수정
